fix: minor fix in GasStationRepository

+++ src/main/java/it/polito/ezgas/repository/GasStationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.polito.ezgas.entity.GasStation;
 
-public interface GasStationRepository extends JpaRepository<GasStation, Long>{
+public interface GasStationRepository extends JpaRepository<GasStation, Integer>{
 
 }
